layers: Split memory related state objects into separate files
Move Image, Buffer and DeviceMemory state objects into their own files
so that it they are easier to deal with.
