+++ BUILD.gn
@@ -91,6 +91,12 @@ core_validation_sources = [
   "layers/core_validation_error_enums.h",
   "layers/core_validation_types.h",
   "layers/base_node.h",
+  "layers/device_memory_state.h",
+  "layers/device_memory_state.cpp",
+  "layers/buffer_state.h",
+  "layers/image_state.h",
+  "layers/image_state.cpp",
+  "layers/ray_tracing_state.h",
   "layers/core_error_location.cpp",
   "layers/core_error_location.h",
   "layers/descriptor_sets.cpp",
