+++ build-android/cmake/layerlib/CMakeLists.txt
@@ -63,6 +63,8 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DVK_USE_PLATFORM_ANDROID_KHR \
                      -fvisibility=hidden")
 add_library(VkLayer_khronos_validation SHARED
         ${SRC_DIR}/layers/state_tracker.cpp
+        ${SRC_DIR}/layers/device_memory_state.cpp
+        ${SRC_DIR}/layers/image_state.cpp
         ${SRC_DIR}/layers/core_validation.cpp
         ${SRC_DIR}/layers/drawdispatch.cpp
         ${SRC_DIR}/layers/convert_to_renderpass2.cpp
