+++ layers/core_validation_types.h
@@ -41,6 +41,9 @@
 #include "image_layout_map.h"
 #include "command_validation.h"
 #include "base_node.h"
+#include "image_state.h"
+#include "buffer_state.h"
+#include "ray_tracing_state.h"
 
 #include <array>
 #include <functional>
@@ -120,17 +123,6 @@ static inline bool QueueFamilyIsExternal(const uint32_t queue_family_index) {
 // Caution: Section 7.7.4 states that "If the values of srcQueueFamilyIndex and dstQueueFamilyIndex are equal, no ownership transfer
 // is performed, and the barrier operates as if they were both set to VK_QUEUE_FAMILY_IGNORED."; this does not handle that case.
 static inline bool QueueFamilyIsIgnored(uint32_t queue_family_index) { return queue_family_index == VK_QUEUE_FAMILY_IGNORED; }
-
-// Intentionally ignore VulkanTypedHandle::node, it is optional
-inline bool operator==(const VulkanTypedHandle &a, const VulkanTypedHandle &b) NOEXCEPT {
-    return a.handle == b.handle && a.type == b.type;
-}
-namespace std {
-template <>
-struct hash<VulkanTypedHandle> {
-    size_t operator()(VulkanTypedHandle obj) const NOEXCEPT { return hash<uint64_t>()(obj.handle) ^ hash<uint32_t>()(obj.type); }
-};
-}  // namespace std
 
 // Flags describing requirements imposed by the pipeline on a descriptor. These
 // can't be checked at pipeline creation time as they depend on the Image or
@@ -224,245 +216,6 @@ struct DESCRIPTOR_POOL_STATE : BASE_NODE {
     }
 };
 
-struct MemRange {
-    VkDeviceSize offset = 0;
-    VkDeviceSize size = 0;
-};
-
-// Data struct for tracking memory object
-struct DEVICE_MEMORY_STATE : public BASE_NODE {
-    void *object;  // Dispatchable object used to create this memory (device of swapchain)
-    safe_VkMemoryAllocateInfo alloc_info;
-    VulkanTypedHandle dedicated_handle;
-    union {
-        VkBufferCreateInfo buffer;
-        VkImageCreateInfo image;
-    } dedicated_create_info;
-    bool is_export;
-    bool is_import;
-    bool is_import_ahb;   // The VUID check depends on if the imported memory is for AHB
-    bool unprotected;     // can't be used for protected memory
-    bool multi_instance;  // Allocated from MULTI_INSTANCE heap or having more than one deviceMask bit set
-    VkExternalMemoryHandleTypeFlags export_handle_type_flags;
-    VkExternalMemoryHandleTypeFlags import_handle_type_flags;
-    // Images for alias search
-    layer_data::unordered_set<IMAGE_STATE *> bound_images;
-
-    MemRange mapped_range;
-    void *shadow_copy_base;          // Base of layer's allocation for guard band, data, and alignment space
-    void *shadow_copy;               // Pointer to start of guard-band data before mapped region
-    uint64_t shadow_pad_size;        // Size of the guard-band data before and after actual data. It MUST be a
-                                     // multiple of limits.minMemoryMapAlignment
-    void *p_driver_data;             // Pointer to application's actual memory
-    VkDeviceSize fake_base_address;  // To allow a unified view of allocations, useful to Synchronization Validation
-
-    DEVICE_MEMORY_STATE(void *disp_object, const VkDeviceMemory in_mem, const VkMemoryAllocateInfo *p_alloc_info,
-                        uint64_t fake_address)
-        : BASE_NODE(in_mem, kVulkanObjectTypeDeviceMemory),
-          object(disp_object),
-          alloc_info(p_alloc_info),
-          dedicated_handle(),
-          dedicated_create_info{},
-          is_export(false),
-          is_import(false),
-          is_import_ahb(false),
-          unprotected(true),
-          multi_instance(false),
-          export_handle_type_flags(0),
-          import_handle_type_flags(0),
-          mapped_range{},
-          shadow_copy_base(0),
-          shadow_copy(0),
-          shadow_pad_size(0),
-          p_driver_data(0),
-          fake_base_address(fake_address){};
-
-    bool IsDedicatedBuffer() const { return dedicated_handle.type == kVulkanObjectTypeBuffer && dedicated_handle.handle != 0; }
-
-    bool IsDedicatedImage() const { return dedicated_handle.type == kVulkanObjectTypeImage && dedicated_handle.handle != 0; }
-
-    VkDeviceMemory mem() const { return handle_.Cast<VkDeviceMemory>(); }
-
-    virtual ~DEVICE_MEMORY_STATE() { Destroy(); }
-
-    void RemoveParent(BASE_NODE *parent_node) override;
-
-    void Destroy() override;
-
-    const BindingsType& ObjectBindings() const { return parent_nodes_; }
-};
-
-// Generic memory binding struct to track objects bound to objects
-struct MEM_BINDING {
-    std::shared_ptr<DEVICE_MEMORY_STATE> mem_state;
-    VkDeviceSize offset;
-    VkDeviceSize size;
-};
-
-class BUFFER_STATE;
-struct BufferBinding {
-    std::shared_ptr<BUFFER_STATE> buffer_state;
-    VkDeviceSize size;
-    VkDeviceSize offset;
-    VkDeviceSize stride;
-
-    BufferBinding() : buffer_state(), size(0), offset(0), stride(0) {}
-    virtual ~BufferBinding() {}
-
-    virtual void reset() { *this = BufferBinding(); }
-};
-
-struct IndexBufferBinding : BufferBinding {
-    VkIndexType index_type;
-
-    IndexBufferBinding() : BufferBinding(), index_type(static_cast<VkIndexType>(0)) {}
-    virtual ~IndexBufferBinding() {}
-
-    virtual void reset() override { *this = IndexBufferBinding(); }
-};
-
-inline bool operator==(MEM_BINDING a, MEM_BINDING b) NOEXCEPT {
-    return a.mem_state == b.mem_state && a.offset == b.offset && a.size == b.size;
-}
-
-namespace std {
-template <>
-struct hash<MEM_BINDING> {
-    size_t operator()(MEM_BINDING mb) const NOEXCEPT {
-        auto intermediate = hash<uint64_t>()(reinterpret_cast<uint64_t &>(mb.mem_state)) ^ hash<uint64_t>()(mb.offset);
-        return intermediate ^ hash<uint64_t>()(mb.size);
-    }
-};
-}  // namespace std
-
-// Superclass for bindable object state (currently images and buffers)
-class BINDABLE : public BASE_NODE {
-  public:
-    using BoundMemorySet = small_unordered_set<DEVICE_MEMORY_STATE *, 1>;
-
-    bool sparse;  // Is this object being bound with sparse memory or not?
-    // Non-sparse binding data
-    MEM_BINDING binding;
-    // Memory requirements for this BINDABLE
-    VkMemoryRequirements requirements;
-    // bool to track if memory requirements were checked
-    bool memory_requirements_checked;
-    // Tracks external memory types creating resource
-    VkExternalMemoryHandleTypeFlags external_memory_handle;
-    // Sparse binding data, initially just tracking MEM_BINDING per mem object
-    //  There's more data for sparse bindings so need better long-term solution
-    // TODO : Need to update solution to track all sparse binding data
-    layer_data::unordered_set<MEM_BINDING> sparse_bindings;
-    // True if memory will be imported/exported from/to an Android Hardware Buffer
-    bool external_ahb;
-    bool unprotected;  // can't be used for protected memory
-
-    BoundMemorySet bound_memory_set_;
-
-    template <typename Handle>
-    BINDABLE(Handle h, VulkanObjectType t)
-        : BASE_NODE(h, t),
-          sparse(false),
-          binding{},
-          requirements{},
-          memory_requirements_checked(false),
-          external_memory_handle(0),
-          sparse_bindings{},
-          external_ahb(false),
-          unprotected(true),
-          bound_memory_set_{} {};
-
-    virtual ~BINDABLE() {
-        if (!Destroyed()) {
-            Destroy();
-        }
-    }
-
-    void Destroy() override;
-
-    // Update the cached set of memory bindings.
-    // Code that changes binding.mem or sparse_bindings must call UpdateBoundMemorySet()
-    void UpdateBoundMemorySet() {
-        bound_memory_set_.clear();
-        if (!sparse) {
-            if (binding.mem_state) bound_memory_set_.insert(binding.mem_state.get());
-        } else {
-            for (const auto &sb : sparse_bindings) {
-                bound_memory_set_.insert(sb.mem_state.get());
-            }
-        }
-    }
-
-    // Return unordered set of memory objects that are bound
-    // Instead of creating a set from scratch each query, return the cached one
-    const BoundMemorySet &GetBoundMemory() const { return bound_memory_set_; }
-
-    void SetMemBinding(std::shared_ptr<DEVICE_MEMORY_STATE> &mem, VkDeviceSize memory_offset);
-    void SetSparseMemBinding(std::shared_ptr<DEVICE_MEMORY_STATE> &mem, const VkDeviceSize mem_offset, const VkDeviceSize mem_size);
-};
-
-class BUFFER_STATE : public BINDABLE {
-  public:
-    VkBufferCreateInfo createInfo;
-    VkDeviceAddress deviceAddress;
-    BUFFER_STATE(VkBuffer buff, const VkBufferCreateInfo *pCreateInfo)
-        : BINDABLE(buff, kVulkanObjectTypeBuffer), createInfo(*pCreateInfo) {
-        if ((createInfo.sharingMode == VK_SHARING_MODE_CONCURRENT) && (createInfo.queueFamilyIndexCount > 0)) {
-            uint32_t *pQueueFamilyIndices = new uint32_t[createInfo.queueFamilyIndexCount];
-            for (uint32_t i = 0; i < createInfo.queueFamilyIndexCount; i++) {
-                pQueueFamilyIndices[i] = pCreateInfo->pQueueFamilyIndices[i];
-            }
-            createInfo.pQueueFamilyIndices = pQueueFamilyIndices;
-        }
-
-        if (createInfo.flags & VK_BUFFER_CREATE_SPARSE_BINDING_BIT) {
-            sparse = true;
-        }
-
-        auto *externalMemoryInfo = LvlFindInChain<VkExternalMemoryBufferCreateInfo>(pCreateInfo->pNext);
-        if (externalMemoryInfo) {
-            external_memory_handle = externalMemoryInfo->handleTypes;
-        }
-    };
-
-    BUFFER_STATE(BUFFER_STATE const &rh_obj) = delete;
-
-    VkBuffer buffer() const { return handle_.Cast<VkBuffer>(); }
-
-    ~BUFFER_STATE() {
-        if ((createInfo.sharingMode == VK_SHARING_MODE_CONCURRENT) && (createInfo.queueFamilyIndexCount > 0)) {
-            delete[] createInfo.pQueueFamilyIndices;
-            createInfo.pQueueFamilyIndices = nullptr;
-        }
-        Destroy();
-    };
-};
-
-class BUFFER_VIEW_STATE : public BASE_NODE {
-  public:
-    VkBufferViewCreateInfo create_info;
-    std::shared_ptr<BUFFER_STATE> buffer_state;
-    VkFormatFeatureFlags format_features;
-    BUFFER_VIEW_STATE(const std::shared_ptr<BUFFER_STATE> &bf, VkBufferView bv, const VkBufferViewCreateInfo *ci)
-        : BASE_NODE(bv, kVulkanObjectTypeBufferView), create_info(*ci), buffer_state(bf) {
-        if (buffer_state) {
-            buffer_state->AddParent(this);
-        }
-    }
-    virtual ~BUFFER_VIEW_STATE() { Destroy(); }
-
-    BUFFER_VIEW_STATE(const BUFFER_VIEW_STATE &rh_obj) = delete;
-
-    VkBufferView buffer_view() const { return handle_.Cast<VkBufferView>(); }
-
-    void Destroy() override {
-        if (buffer_state) {
-            buffer_state->RemoveParent(this);
-        }
-        BASE_NODE::Destroy();
-    }
-};
-
 struct SAMPLER_STATE : public BASE_NODE {
     VkSamplerCreateInfo createInfo;
     VkSamplerYcbcrConversion samplerConversion = VK_NULL_HANDLE;
@@ -477,205 +230,6 @@ struct SAMPLER_STATE : public BASE_NODE {
     }
 
     VkSampler sampler() const { return handle_.Cast<VkSampler>(); }
-};
-
-class IMAGE_STATE : public BINDABLE {
-  public:
-    safe_VkImageCreateInfo safe_create_info;
-    VkImageCreateInfo &createInfo;
-    bool valid;               // If this is a swapchain image backing memory track valid here as it doesn't have DEVICE_MEMORY_STATE
-    bool acquired;            // If this is a swapchain image, has it been acquired by the app.
-    bool shared_presentable;  // True for a front-buffered swapchain image
-    bool layout_locked;       // A front-buffered image that has been presented can never have layout transitioned
-    bool get_sparse_reqs_called;         // Track if GetImageSparseMemoryRequirements() has been called for this image
-    bool sparse_metadata_required;       // Track if sparse metadata aspect is required for this image
-    bool sparse_metadata_bound;          // Track if sparse metadata aspect is bound to this image
-    bool has_ahb_format;                 // True if image was created with an external Android format
-    bool is_swapchain_image;             // True if image is a swapchain image
-    uint64_t ahb_format;                 // External Android format, if provided
-    VkImageSubresourceRange full_range;  // The normalized ISR for all levels, layers (slices), and aspects
-    VkSwapchainKHR create_from_swapchain;
-    VkSwapchainKHR bind_swapchain;
-    uint32_t bind_swapchain_imageIndex;
-    image_layout_map::Encoder range_encoder;
-    VkFormatFeatureFlags format_features = 0;
-    // Need to memory requirments for each plane if image is disjoint
-    bool disjoint;  // True if image was created with VK_IMAGE_CREATE_DISJOINT_BIT
-    VkMemoryRequirements plane0_requirements;
-    bool plane0_memory_requirements_checked;
-    VkMemoryRequirements plane1_requirements;
-    bool plane1_memory_requirements_checked;
-    VkMemoryRequirements plane2_requirements;
-    bool plane2_memory_requirements_checked;
-
-    const image_layout_map::Encoder subresource_encoder;                             // Subresource resolution encoder
-    std::unique_ptr<const subresource_adapter::ImageRangeEncoder> fragment_encoder;  // Fragment resolution encoder
-    const VkDevice store_device_as_workaround;                                       // TODO REMOVE WHEN encoder can be const
-    VkDeviceSize swapchain_fake_address;  // Needed for swapchain syncval, since there is no VkDeviceMemory::fake_base_address
-
-    std::vector<VkSparseImageMemoryRequirements> sparse_requirements;
-    IMAGE_STATE(VkDevice dev, VkImage img, const VkImageCreateInfo *pCreateInfo);
-    IMAGE_STATE(IMAGE_STATE const &rh_obj) = delete;
-
-    VkImage image() const { return handle_.Cast<VkImage>(); }
-
-    layer_data::unordered_set<IMAGE_STATE *> aliasing_images;
-    bool IsCompatibleAliasing(IMAGE_STATE *other_image_state) const;
-
-    bool IsCreateInfoEqual(const VkImageCreateInfo &other_createInfo) const;
-    bool IsCreateInfoDedicatedAllocationImageAliasingCompatible(const VkImageCreateInfo &other_createInfo) const;
-
-    inline bool IsImageTypeEqual(const VkImageCreateInfo &other_createInfo) const {
-        return createInfo.imageType == other_createInfo.imageType;
-    }
-    inline bool IsFormatEqual(const VkImageCreateInfo &other_createInfo) const {
-        return createInfo.format == other_createInfo.format;
-    }
-    inline bool IsMipLevelsEqual(const VkImageCreateInfo &other_createInfo) const {
-        return createInfo.mipLevels == other_createInfo.mipLevels;
-    }
-    inline bool IsUsageEqual(const VkImageCreateInfo &other_createInfo) const { return createInfo.usage == other_createInfo.usage; }
-    inline bool IsSamplesEqual(const VkImageCreateInfo &other_createInfo) const {
-        return createInfo.samples == other_createInfo.samples;
-    }
-    inline bool IsTilingEqual(const VkImageCreateInfo &other_createInfo) const {
-        return createInfo.tiling == other_createInfo.tiling;
-    }
-    inline bool IsArrayLayersEqual(const VkImageCreateInfo &other_createInfo) const {
-        return createInfo.arrayLayers == other_createInfo.arrayLayers;
-    }
-    inline bool IsInitialLayoutEqual(const VkImageCreateInfo &other_createInfo) const {
-        return createInfo.initialLayout == other_createInfo.initialLayout;
-    }
-    inline bool IsSharingModeEqual(const VkImageCreateInfo &other_createInfo) const {
-        return createInfo.sharingMode == other_createInfo.sharingMode;
-    }
-    inline bool IsExtentEqual(const VkImageCreateInfo &other_createInfo) const {
-        return (createInfo.extent.width == other_createInfo.extent.width) &&
-               (createInfo.extent.height == other_createInfo.extent.height) &&
-               (createInfo.extent.depth == other_createInfo.extent.depth);
-    }
-    inline bool IsQueueFamilyIndicesEqual(const VkImageCreateInfo &other_createInfo) const {
-        return (createInfo.queueFamilyIndexCount == other_createInfo.queueFamilyIndexCount) &&
-               (createInfo.queueFamilyIndexCount == 0 ||
-                memcmp(createInfo.pQueueFamilyIndices, other_createInfo.pQueueFamilyIndices,
-                       createInfo.queueFamilyIndexCount * sizeof(createInfo.pQueueFamilyIndices[0])) == 0);
-    }
-
-    ~IMAGE_STATE() {
-        if ((createInfo.sharingMode == VK_SHARING_MODE_CONCURRENT) && (createInfo.queueFamilyIndexCount > 0)) {
-            delete[] createInfo.pQueueFamilyIndices;
-            createInfo.pQueueFamilyIndices = nullptr;
-        }
-        Destroy();
-    };
-
-    void Destroy() override {
-        RemoveAliasingImages();
-        BINDABLE::Destroy();
-    }
-
-    void AddAliasingImage(layer_data::unordered_set<IMAGE_STATE *> &bound_images);
-    void RemoveAliasingImages();
-};
-
-class IMAGE_VIEW_STATE : public BASE_NODE {
-  public:
-    VkImageViewCreateInfo create_info;
-    const VkImageSubresourceRange normalized_subresource_range;
-    const image_layout_map::RangeGenerator range_generator;
-    VkSampleCountFlagBits samples;
-    unsigned descriptor_format_bits;
-    VkSamplerYcbcrConversion samplerConversion;  // Handle of the ycbcr sampler conversion the image was created with, if any
-    VkFilterCubicImageViewImageFormatPropertiesEXT filter_cubic_props;
-    VkFormatFeatureFlags format_features;
-    VkImageUsageFlags inherited_usage;  // from spec #resources-image-inherited-usage
-    std::shared_ptr<IMAGE_STATE> image_state;
-    IMAGE_VIEW_STATE(const std::shared_ptr<IMAGE_STATE> &image_state, VkImageView iv, const VkImageViewCreateInfo *ci);
-    IMAGE_VIEW_STATE(const IMAGE_VIEW_STATE &rh_obj) = delete;
-
-    VkImageView image_view() const { return handle_.Cast<VkImageView>(); }
-
-    virtual ~IMAGE_VIEW_STATE() { Destroy(); }
-
-    bool OverlapSubresource(const IMAGE_VIEW_STATE &compare_view) const;
-
-    void Destroy() override {
-        if (image_state) {
-            image_state->RemoveParent(this);
-        }
-        BASE_NODE::Destroy();
-    }
-};
-
-class ACCELERATION_STRUCTURE_STATE : public BINDABLE {
-  public:
-    safe_VkAccelerationStructureCreateInfoNV create_infoNV = {};
-    safe_VkAccelerationStructureInfoNV build_info;
-    bool memory_requirements_checked = false;
-    VkMemoryRequirements2 memory_requirements;
-    bool build_scratch_memory_requirements_checked = false;
-    VkMemoryRequirements2 build_scratch_memory_requirements;
-    bool update_scratch_memory_requirements_checked = false;
-    VkMemoryRequirements2 update_scratch_memory_requirements;
-    bool built = false;
-    uint64_t opaque_handle = 0;
-    const VkAllocationCallbacks *allocator = NULL;
-    ACCELERATION_STRUCTURE_STATE(VkAccelerationStructureNV as, const VkAccelerationStructureCreateInfoNV *ci)
-        : BINDABLE(as, kVulkanObjectTypeAccelerationStructureNV),
-          create_infoNV(ci),
-          memory_requirements{},
-          build_scratch_memory_requirements_checked{},
-          build_scratch_memory_requirements{},
-          update_scratch_memory_requirements_checked{},
-          update_scratch_memory_requirements{} {}
-    ACCELERATION_STRUCTURE_STATE(const ACCELERATION_STRUCTURE_STATE &rh_obj) = delete;
-
-    VkAccelerationStructureNV acceleration_structure() const { return handle_.Cast<VkAccelerationStructureNV>(); }
-};
-
-class ACCELERATION_STRUCTURE_STATE_KHR : public BINDABLE {
-  public:
-    safe_VkAccelerationStructureCreateInfoKHR create_infoKHR = {};
-    safe_VkAccelerationStructureBuildGeometryInfoKHR build_info_khr;
-    bool memory_requirements_checked = false;
-    VkMemoryRequirements2 memory_requirements;
-    bool build_scratch_memory_requirements_checked = false;
-    VkMemoryRequirements2 build_scratch_memory_requirements;
-    bool update_scratch_memory_requirements_checked = false;
-    VkMemoryRequirements2 update_scratch_memory_requirements;
-    bool built = false;
-    uint64_t opaque_handle = 0;
-    const VkAllocationCallbacks *allocator = NULL;
-    ACCELERATION_STRUCTURE_STATE_KHR(VkAccelerationStructureKHR as, const VkAccelerationStructureCreateInfoKHR *ci)
-        : BINDABLE(as, kVulkanObjectTypeAccelerationStructureKHR),
-          create_infoKHR(ci),
-          memory_requirements{},
-          build_scratch_memory_requirements_checked{},
-          build_scratch_memory_requirements{},
-          update_scratch_memory_requirements_checked{},
-          update_scratch_memory_requirements{} {}
-    ACCELERATION_STRUCTURE_STATE_KHR(const ACCELERATION_STRUCTURE_STATE_KHR &rh_obj) = delete;
-
-    VkAccelerationStructureKHR acceleration_structure() const { return handle_.Cast<VkAccelerationStructureKHR>(); }
-};
-
-struct SWAPCHAIN_IMAGE {
-    IMAGE_STATE *image_state = nullptr;
-    layer_data::unordered_set<IMAGE_STATE *> bound_images;
-};
-
-class SWAPCHAIN_NODE : public BASE_NODE {
-  public:
-    safe_VkSwapchainCreateInfoKHR createInfo;
-    std::vector<SWAPCHAIN_IMAGE> images;
-    bool retired = false;
-    bool shared_presentable = false;
-    uint32_t get_swapchain_image_count = 0;
-    SWAPCHAIN_NODE(const VkSwapchainCreateInfoKHR *pCreateInfo, VkSwapchainKHR swapchain)
-        : BASE_NODE(swapchain, kVulkanObjectTypeSwapchainKHR), createInfo(pCreateInfo) {}
-
-    VkSwapchainKHR swapchain() const { return handle_.Cast<VkSwapchainKHR>(); }
 };
 
 // Store the DAG.
@@ -880,11 +434,6 @@ struct CBVertexBufferBindingInfo {
     std::vector<BufferBinding> vertex_buffer_bindings;
 };
 
-static inline bool operator==(const VkImageSubresource &lhs, const VkImageSubresource &rhs) {
-    bool is_equal = (lhs.aspectMask == rhs.aspectMask) && (lhs.mipLevel == rhs.mipLevel) && (lhs.arrayLayer == rhs.arrayLayer);
-    return is_equal;
-}
-
 // Canonical dictionary for PushConstantRanges
 using PushConstantRangesDict = hash_util::Dictionary<PushConstantRanges>;
 using PushConstantRangesId = PushConstantRangesDict::Id;
@@ -963,60 +512,6 @@ struct interface_var {
           is_atomic_operation(false),
           is_sampler_implicitLod_dref_proj(false),
           is_sampler_bias_offset(false) {}
-};
-
-// Safe struct that spans NV and KHR VkRayTracingPipelineCreateInfo structures.
-// It is a safe_VkRayTracingPipelineCreateInfoKHR and supports construction from
-// a VkRayTracingPipelineCreateInfoNV.
-class safe_VkRayTracingPipelineCreateInfoCommon : public safe_VkRayTracingPipelineCreateInfoKHR {
-  public:
-    safe_VkRayTracingPipelineCreateInfoCommon() : safe_VkRayTracingPipelineCreateInfoKHR() {}
-    safe_VkRayTracingPipelineCreateInfoCommon(const VkRayTracingPipelineCreateInfoNV *pCreateInfo)
-        : safe_VkRayTracingPipelineCreateInfoKHR() {
-        initialize(pCreateInfo);
-    }
-    void initialize(const VkRayTracingPipelineCreateInfoNV *pCreateInfo) {
-        safe_VkRayTracingPipelineCreateInfoNV nvStruct;
-        nvStruct.initialize(pCreateInfo);
-
-        sType = nvStruct.sType;
-
-        // Take ownership of the pointer and null it out in nvStruct
-        pNext = nvStruct.pNext;
-        nvStruct.pNext = nullptr;
-
-        flags = nvStruct.flags;
-        stageCount = nvStruct.stageCount;
-
-        pStages = nvStruct.pStages;
-        nvStruct.pStages = nullptr;
-
-        groupCount = nvStruct.groupCount;
-        maxRecursionDepth = nvStruct.maxRecursionDepth;
-        layout = nvStruct.layout;
-        basePipelineHandle = nvStruct.basePipelineHandle;
-        basePipelineIndex = nvStruct.basePipelineIndex;
-
-        assert(pGroups == nullptr);
-        if (nvStruct.groupCount && nvStruct.pGroups) {
-            pGroups = new safe_VkRayTracingShaderGroupCreateInfoKHR[groupCount];
-            for (uint32_t i = 0; i < groupCount; ++i) {
-                pGroups[i].sType = nvStruct.pGroups[i].sType;
-                pGroups[i].pNext = nvStruct.pGroups[i].pNext;
-                pGroups[i].type = nvStruct.pGroups[i].type;
-                pGroups[i].generalShader = nvStruct.pGroups[i].generalShader;
-                pGroups[i].closestHitShader = nvStruct.pGroups[i].closestHitShader;
-                pGroups[i].anyHitShader = nvStruct.pGroups[i].anyHitShader;
-                pGroups[i].intersectionShader = nvStruct.pGroups[i].intersectionShader;
-                pGroups[i].intersectionShader = nvStruct.pGroups[i].intersectionShader;
-                pGroups[i].pShaderGroupCaptureReplayHandle = nullptr;
-            }
-        }
-    }
-    void initialize(const VkRayTracingPipelineCreateInfoKHR *pCreateInfo) {
-        safe_VkRayTracingPipelineCreateInfoKHR::initialize(pCreateInfo);
-    }
-    uint32_t maxRecursionDepth;  // NV specific
 };
 
 struct SHADER_MODULE_STATE;
